# docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  ssrtheme.InstallThemeRequest:
    properties:
      downloadUrl:
        type: string
      marketId:
        type: integer
      themeName:
        type: string
      version:
        type: string
    required:
    - downloadUrl
    - themeName
    type: object
  ssrtheme.StartThemeRequest:
    properties:
      port:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/ssr-theme/{name}:
    delete:
      description: 卸载指定的 SSR 主题
      parameters:
      - description: 主题名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 卸载 SSR 主题
      tags:
      - SSR主题管理
  /api/admin/ssr-theme/{name}/start:
    post:
      consumes:
      - application/json
      description: 启动指定的 SSR 主题，并将其设为当前主题
      parameters:
      - description: 主题名称
        in: path
        name: name
        required: true
        type: string
      - description: 启动参数
        in: body
        name: request
        schema:
          $ref: '#/definitions/ssrtheme.StartThemeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 启动 SSR 主题
      tags:
      - SSR主题管理
  /api/admin/ssr-theme/{name}/status:
    get:
      description: 获取指定 SSR 主题的状态信息
      parameters:
      - description: 主题名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取 SSR 主题状态
      tags:
      - SSR主题管理
  /api/admin/ssr-theme/{name}/stop:
    post:
      description: 停止指定的 SSR 主题
      parameters:
      - description: 主题名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 停止 SSR 主题
      tags:
      - SSR主题管理
  /api/admin/ssr-theme/install:
    post:
      consumes:
      - application/json
      description: 从指定 URL 下载并安装 SSR 主题
      parameters:
      - description: 安装请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ssrtheme.InstallThemeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 安装 SSR 主题
      tags:
      - SSR主题管理
...
swagger: "2.0"